+++ src/webapp/components/predictor-edit-wizard/PredictorEditWizard.tsx
@@ -6,9 +6,9 @@ import { ArrowBack, ArrowForward } from "@material-ui/icons";
 import React, { useCallback } from "react";
 import { Form } from "react-final-form";
 import styled from "styled-components";
-import { Predictor } from "../../domain/entities/Predictor";
-import { StyledForm } from "../components/form/StyledForm";
-import { useQueryState } from "../hooks/useQueryState";
+import { Predictor } from "../../../domain/entities/Predictor";
+import { StyledForm } from "../../components/form/StyledForm";
+import { useQueryState } from "../../hooks/useQueryState";
 import { GeneralInfoStep } from "./steps/GeneralInfoStep";
 import { GeneratorStep } from "./steps/GeneratorStep";
 import { SampleStep } from "./steps/SampleStep";
